% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_liv.R
\name{.get_test_info}
\alias{.get_test_info}
\title{Get test info from the filename}
\usage{
.get_test_info(file)
}
\arguments{
\item{file}{Path to an LIV data file.}
}
\value{
A single row data frame
}
\description{
Get test info from the filename
}
