% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_osa.R
\name{.osa_get_test_info}
\alias{.osa_get_test_info}
\title{Get test info from the filename}
\usage{
.osa_get_test_info(file)
}
\arguments{
\item{file}{Path to an LIV data file.}
}
\value{
A single row data frame
}
\description{
Get test info from the filename
}
