% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{eload}
\alias{eload}
\title{Loader of mapping of names to values into variables within an environment}
\usage{
eload(loadDat, loadEnvir = loadrEnv(), preserve = FALSE)
}
\arguments{
\item{loadDat}{Collection of bindings between name and value, e.g. a 
\code{list} or \code{environment}}

\item{loadEnvir}{Name (character string) for the environment to create or
update.}

\item{preserve}{Whether to retain the value for an already-bound name.}
}
\description{
\code{eload} takes a collection of bindings between name and value and 
uses those bindings to create or update an environment. A value for 
preserved according to the argument to \code{preserve}. If you want to load
directly into the current env, look at \code{list2env} with
\code{environment()}
}
