% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{getLoadEnvir}
\alias{getLoadEnvir}
\title{A function used by eload() to create the global shared variable
repository if it doesn't exist, or return it if it does.
Internal function.}
\usage{
getLoadEnvir(loadEnvir = loadrEnv())
}
\description{
A function used by eload() to create the global shared variable
repository if it doesn't exist, or return it if it does.
Internal function.
}
