% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dependencies.R
\name{install_packages_if_necessary}
\alias{install_packages_if_necessary}
\title{Selectively install the packages you specify}
\usage{
install_packages_if_necessary(dependencies,
  installed_packages = utils::installed.packages())
}
\arguments{
\item{dependencies}{A character vector of packages your code requires. If you
want to depend on a CRAN package, just use its bare name: \code{"ggplot"}.
If you want to depend on a package hosted on Github, specify your package
as \code{"user/repo"} and loadr will automatically install the package from
Github.}

\item{installed_packages}{A character vector of packages already installed.
loadr will not attempt to install these packages. loadr will only install
packages in \code{dependencies} that do not appear here. The parameter
defaults to \code{utils::installed.packages()}, which I'm pretty sure is
what you want. But on the off chance you want to supply some alternative
list of packages, you can just override this default.}
}
\value{
An invisible character vector of the packages that were required but
  not present in \code{installed_packages}. Ideally, this should be a vector
  of just the packages that were installed, but on the off chance an
  installation fails, a package name may appear here even if it wasn't
  successfully installed. So, I'll try to work on that.
}
\description{
This function takes a vector of packages, determines which packages aren't
currently installed, and installs only the missing dependencies.
}
\examples{
\dontrun{
install_packages_if_necessary("ggplot2")
# specifying a package as "user/repo" will
# install from Github using devtools::install_github()
install_packages_if_necessary("briandk/granovagg")
my_dependencies = c("tidyr", "hadley/testthat")
install_packages_if_necessary(my_dependencies)
}

}

