% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LAMfit0}
\alias{LAMfit0}
\title{Function to be minimised within \code{LAM0}}
\usage{
LAMfit0(data, pars)
}
\arguments{
\item{data}{data.frame of Q and Co values input to \code{LAM} or
\code{LAMnls}}

\item{pars}{starting parameters}
}
\description{
Used internally by \code{LAM0}
}
\details{
Shouldn't be used directly.
}

