% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LAMnlplot}
\alias{LAMnlplot}
\title{Plot "LAMnl" object}
\usage{
LAMnlplot(x, f.unit = "cumecs", c.unit = expression("ugl"^-1), ...)
}
\arguments{
\item{x}{a "LAMnl" object}

\item{f.unit}{unit of flow as a text string. Used in labelling of plots.}

\item{c.unit}{unit of concentration as a text string. Used in labelling of
plots.}

\item{...}{further arguments passed to and from other methods}
}
\description{
Plots a four panel graphic summairsing Load Apportionment Model fits.
}
\details{
\code{y} is included for consistency with generic \code{print} but is
     not used
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMwensumnlxb <- LAMnlxb(Q, Co)

     LAMnlplot(LAMwensumnlxb)
}

