% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LAMnls}
\alias{LAMnls}
\title{Non-linear lest squares Load Apportionment Model}
\usage{
LAMnls(Q, Co, pars = c(A = 1, B = 1, C = 1, D = 1), b.upper = 1,
  method = "port")
}
\arguments{
\item{Q}{A vector or single element data frame or matrix. Discharge at time
of \code{Co}. Must be same length as Co}

\item{Co}{A vector or single element data frame or matrix. Concentration at
time of \code{Q}. Must be same length as Q.}

\item{pars}{A list or vector of length 4. Provides starting values for the
four optimisation parameters in the order A, B, C, D}

\item{b.upper}{Numeric. Upper limit of B.}

\item{method}{Character string. "port" or "plinear" See \code{?nls} for more
     info on each}
}
\value{
The function \code{summary} can be used to obtain and print a
     concise object of important outputs. The generic fucntions
     \code{fitted.values} and \code{resid} can be used on LAM objects.
     An object of class "LAM" consisting of the following components.
     \itemize{
             \item \code{nls.summary.out} standard output from
             \code{summary.nls}
             \item \code{pars} parameter estimates
             \item \code{fitted.values} fitted values
             \item \code{residuals} residuals
             \item \code{diffuse.point} list of 2 vectors; diffuse and
             point source contributions to \code{fitted.values}
             \item \code{raw} list of 2 vectors; input Q and Co
             \item \code{xover} value of Q where diffuse contributions first
             exceed point
             \item \code{ps.time} percentage of samples (or time in some
             cirumstances) where point source inputs are greater than diffuse
              source inputs
     }
}
\description{
\code{LAMnls} Returns a LAM object by optimisation across A, B, C and D
     paramemeters using \code{nls}
}
\details{
This is an R implementation of the Load Apportionment Model by Bowes et
     al.

     The Load Apportionment Model attributes load at flow x between
     point and diffuse sources based on the relationship between flow and
     concentration. The model is described as
     \eqn{Co = A . Q ^ B-1 + C . Q ^ D-1}. Where Co is concentration, Q is
     discharge and A, B, C and D are parameters to be estimated empiraically.

     This version uses \code{nls} to perform the optimisation. For the
     \code{optim} version use \code{LAM}. Optimisation is performed over
     four parameters; A, B, C and D. By default B is restircted to be 1.
     \code{LAMnls} attempts to solve the Load Apportionment Model using user
     supplied (or default) parameter starting values. If no solution can be
     found the equation is solved using \code{optim}. The output parameters
     are then used as the stating values for \code{nls}. Two methods for
     \code{nls} can be used. "port" allows for upper and lower parameter
     values to be set, "plinear" does not. More methods will be added later.
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMwensumnls <- LAMnls(Q, Co)

     LAMwensumnls$pars ## get parameters

     ## specift starting values
     LAMwensumnls2 <- LAMnls(Q, Co, pars = c(A= 50, B = 1, C = 300, D = 0))
     LAMnlsplot(LAMwensumnls2)
}
\references{
Bowes MJ, Smith JT, Javie H, Neal C (2008). Modelling of
     phosphorus inputs to rivers from diffuse and point sources. Science
     of the Total Environment 395. pp125-138.
}

