% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LAMnlsplot}
\alias{LAMnlsplot}
\title{Plot "LAMnls" object}
\usage{
LAMnlsplot(x, unit = "Ml/d", ...)
}
\arguments{
\item{x}{a "LAMnls" object}

\item{unit}{unit of flow as a text string. Used in labelling of plots.}

\item{...}{further arguments passed to and from other methods}
}
\description{
Plots a four panel graphic summairsing Load Apportionment Model fits.
}
\details{
\code{y} is included for consistency with generic \code{print} but is
     not used
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMwensumnls <- LAMnls(Q, Co)

     LAMnlsplot(LAMwensumnls)
}

