% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LAMout}
\alias{LAMout}
\title{Produce fitted LAM values}
\usage{
LAMout(data, pars)
}
\arguments{
\item{data}{data.frame of Q and Co values input to \code{LAM} or
\code{LAMnls}}

\item{pars}{estiamted parameters}
}
\value{
\itemize{
             \item \code{Fpred} fitted values
             \item \code{Fp} point source contribution to \code{Fpred}
             \item \code{Fd} diffuse source contribution to \code{Fpred}
             \item \code{Fp.load} instantaneous point source load. Units
             are arbitary and dependent on input units
             \item \code{Fd.load} instantaneous diffuse source load. Units
             are arbitary and dependent on input units
             \item \code{Fp.load.day} Daily point source load. Units
             are arbitary and dependent on input units
             \item \code{Fd.load.day} Daily diffuse source load. Units
             are arbitary and dependent on input units
     }
}
\description{
Used internally by \code{LAM} and \code{LAMnls} to produce fitted values
     and diffuse and ponit soruce contributions at each point.
}
\details{
Shouldn't be used directly.
}

