% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library2.R
\name{library2}
\alias{library2}
\title{Attach package}
\usage{
library2(x, ..., quiet = TRUE)
}
\arguments{
\item{x}{a string indicating the name of the package to attach.}

\item{...}{objects passed to \code{library} or \code{pak::pkg_install}.}

\item{quiet}{If \code{FALSE}, the function prints output. Defaults to \code{TRUE}}
}
\value{
NULL invisibly.
}
\description{
Attach package
}
