% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_build_log.R
\name{load_build_log}
\alias{load_build_log}
\title{Load the ELS G2 Build List}
\usage{
load_build_log(
  path =
    "/Users/brian/Library/CloudStorage/Dropbox-POETTechnologies/Brian Pile/Brian Pile/Documents/ELS Gen2 project repository/test and measurement/assy info/PT11B build log.xlsx"
)
}
\arguments{
\item{path}{A path to the build log}
}
\value{
A data frame
}
\description{
Load the ELS G2 Build List
}
