% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_iv.R
\name{load_iv}
\alias{load_iv}
\title{Load IV files}
\usage{
load_iv(file)
}
\arguments{
\item{file}{Path to the IV file}
}
\value{
A data frame
}
\description{
Load IV files
}
\examples{
filepath <- system.file("extdata", "25FC01497_CH1.csv", package = "loadr")
if (file.exists(filepath)) {
  df <- load_iv(filepath)
  head(df)
}
}
