% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_liv.R
\name{load_liv}
\alias{load_liv}
\title{Load LIV files}
\usage{
load_liv(file)
}
\arguments{
\item{file}{Path to an LIV data file.}
}
\value{
A data frame
}
\description{
Load LIV files
}
\examples{
filepath <- system.file("extdata",
   "DT7_25FC01497-CF5_B4_C6-CH1-Front-Pad-Stage_500mA_50C_NAV_00_LIV.xlsx",
   package = "loadr")
if (file.exists(filepath)) {
  df <- load_liv(filepath)
  head(df)
}
}
