% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_osa.R
\name{load_osa}
\alias{load_osa}
\title{Load OSA files}
\usage{
load_osa(file)
}
\arguments{
\item{file}{Path to the OSA file}
}
\value{
A data frame
}
\description{
Load OSA files
}
\examples{
filepath <- system.file("extdata",
   "DT7_25FC01497-CF5_B4_C6-CH1-WG-Pad-Stage_100mA_50C_NAV_00_OSA.xlsx",
   package = "loadr")
if (file.exists(filepath)) {
  df <- load_osa(filepath)
  head(df)
}
}
