% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dependencies.R
\name{load_required_packages}
\alias{load_required_packages}
\title{Load packages you need; optionally install them if they're not already
installed}
\usage{
load_required_packages(required_packages,
  try_installing_if_not_present = TRUE)
}
\arguments{
\item{required_packages}{A character vector of the quoted names of packages
you'd like loaded via the \code{library()} function. Because
\code{library()} isn't vectorised, this function lets you avoid typing a
separate \code{library()} call for each package you want loaded.}

\item{try_installing_if_not_present}{Should the function try to install any
of your dependencies if need be? This setting is useful if you know what
your code requires, but you can't safely assume which (if any) of your
required packages will be installed on a user's system. Defaults to
\code{TRUE}}
}
\value{
an invisible copy of the dependencies you passed in
}
\description{
Load packages you need; optionally install them if they're not already
installed
}
\examples{
load_required_packages(c("boot", "MASS"))

# If you wish to be super-careful,
# you can say NO to automatic package installation.
# If you say no, the function will only try to load your requirements;
# it won't attempt to install anything new even if the package
# you're requiring isn't currently installed
load_required_packages("utils", try_installing_if_not_present = FALSE)
\dontrun{ #
https://github.com/hadley/tidyverse
some_packages_from_the_tidyverse <- c( "ggplot2"
                                       , "tidyr"
                                       , "broom" )
load_required_packages(some_packages_from_the_tidyverse) }

}

