% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{loadr}
\alias{loadr}
\title{Attach several packages}
\usage{
loadr(..., .x = NULL)
}
\arguments{
\item{...}{package names quoted or not.}

\item{.x}{a vector or a list that is concatenated to \code{...}.}
}
\value{
NULL invisibly.
}
\description{
Attach several packages
}
