% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{loadrEnv}
\alias{loadrEnv}
\title{Sets or getsa global variable specifying the default environment name for
\code{\link{loadr}}.}
\usage{
loadrEnv(envName = NULL)
}
\arguments{
\item{envName}{Name of environment where shared variables should be stored.
Leave NULL to retrieve current environment name.}
}
\description{
Sets or getsa global variable specifying the default environment name for
\code{\link{loadr}}.
}
\examples{
loadrEnv("SV")
}
