% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.LAM}
\alias{print.LAM}
\title{Print "LAM" objects}
\usage{
\method{print}{LAM}(x, ...)
}
\arguments{
\item{x}{an object of class "LAM"}

\item{...}{further arguments passed to and from other methods}
}
\description{
\code{print} method for class "LAM"
}

