% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.LAMnl}
\alias{print.LAMnl}
\title{Print "LAMnl" objects}
\usage{
\method{print}{LAMnl}(x, ...)
}
\arguments{
\item{x}{an object of class "LAMnl"}

\item{...}{further arguments passed to and from other methods}
}
\description{
\code{print} method for class "LAMnl"
}

