% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.LAMnls}
\alias{print.LAMnls}
\title{Print "LAMnls" objects}
\usage{
\method{print}{LAMnls}(x, ...)
}
\arguments{
\item{x}{an object of class "LAMnls"}

\item{...}{further arguments passed to and from other methods}
}
\description{
\code{print} method for class "LAMnls"
}

