% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.LAM}
\alias{summary.LAM}
\title{Summary of "LAM" objects}
\usage{
\method{summary}{LAM}(object, ...)
}
\arguments{
\item{object}{an object of class "LAM"}

\item{...}{further arguments passed to and from other methods}
}
\value{
\itemize{
             \item \code{pars} data frame of parameter estiamtes
             \item \code{SumofSq} Sum of Squared errors
             \item \code{CrossOverFlow} value of Q where diffuse
             contributions firstmexceed point
             \item \code{PointSourceDominatedPercentage} percentage of
             samples (or time in some cirumstances) where point source inputs
             are greater than diffuse source inputs
     }
}
\description{
\code{summary} method for class "LAM"
}
\details{
Currently has no associated \code{print} method
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMeg <- LAM(Q, Co)

     foo <- summary(LAMeg)
     foo
     foo$pars ## get parameters
}

