% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.LAMnl}
\alias{summary.LAMnl}
\title{Summary of "LAMnl" objects}
\usage{
\method{summary}{LAMnl}(object, ...)
}
\arguments{
\item{object}{an object of class "LAMnl"}

\item{...}{further arguments passed to and from other methods}
}
\value{
\itemize{
             \item \code{overview} output from \code{summary.nl}
             \item \code{pars} data frame of parameter estiamtes
             \item \code{SumofSq} Sum of Squared errors
             \item \code{CrossOverFlow} value of Q where diffuse
             contributions firstmexceed point
             \item \code{PointSourceDominatedPercentage} percentage of
             samples (or time in some cirumstances) where point source inputs
             are greater than diffuse source inputs
             \item \code{ps.load} percentage of load derived from point
              sources
     }
}
\description{
\code{summary} method for class "LAMnl"
}
\details{
Currently has no associated \code{print} method
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMnleg <- LAMnlxb(Q, Co)

     foo <- summary(LAMnleg)
     foo
     foo$pars ## get parameters
}

