% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.LAMnls}
\alias{summary.LAMnls}
\title{Summary of "LAMnls" objects}
\usage{
\method{summary}{LAMnls}(object, ...)
}
\arguments{
\item{object}{an object of class "LAMnls"}

\item{...}{further arguments passed to and from other methods}
}
\value{
\itemize{
             \item \code{overview} output from \code{summary.nls}
             \item \code{pars} data frame of parameter estiamtes
             \item \code{SumofSq} Sum of Squared errors
             \item \code{CrossOverFlow} value of Q where diffuse
             contributions firstmexceed point
             \item \code{PointSourceDominatedPercentage} percentage of
             samples (or time in some cirumstances) where point source inputs
             are greater than diffuse source inputs
     }
}
\description{
\code{summary} method for class "LAMnls"
}
\details{
Currently has no associated \code{print} method
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMnlseg <- LAMnls(Q, Co)

     foo <- summary(LAMnlseg)
     foo
     foo$pars ## get parameters
}

