% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{sv}
\alias{sv}
\title{Show shared variables
Gives a list of shared variable contents.}
\usage{
sv(envir = "SV")
}
\arguments{
\item{envir}{Character vector name of the environment to display.}
}
\description{
Show shared variables
Gives a list of shared variable contents.
}
\examples{
sv()
}
