% Generated by roxygen2 (4.0.1): do not edit by hand
\name{till}
\alias{till}
\title{River Till LAM sample data}
\format{A data frame with 248 rows and 3 variables}
\source{
Environment Agency
     \url{https://www.gov.uk/government/organisations/environment-agency}
     databases WISKI for discharge data and WIMS for water qualtiy data.
}
\description{
Discharge and total phosphorus from a single River Till
     sample point.  The variables are as follows:

\itemize{
  \item Date. Date of observation
  \item Q. River discahrge at sample point (cumecs)
  \item TP. Total phosphorus concentration (mg per litre)

}
}

