% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{getLoadEnvir}
\alias{getLoadEnvir}
\title{A function used by eload() to create the global shared variable
environment if it doesn't exist, or return it if it does.}
\usage{
getLoadEnvir(loadEnvir = loadrEnv())
}
\arguments{
\item{loadEnvir}{Name of the environment to get.
Internal function.}
}
\description{
A function used by eload() to create the global shared variable
environment if it doesn't exist, or return it if it does.
}
