% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LAMplot}
\alias{LAMplot}
\title{Plot "LAM" object}
\usage{
LAMplot(x, f.unit = "cumecs", c.unit = expression("ugl"^-1), ...)
}
\arguments{
\item{x}{a "LAM" object}

\item{f.unit}{unit of flow as a text string. Used in labelling of plots.}

\item{c.unit}{unit of concentration as a text string. Used in labelling of
plots.}

\item{...}{further arguments passed to and from other methods}
}
\description{
Plots a four panel graphic summairsing Load Apportionment Model fits.
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMwensum <- LAM(Q, Co)

     LAMplot(LAMwensum)
}

