% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LAMplot}
\alias{LAMplot}
\title{Plot "LAM" object}
\usage{
LAMplot(x, y, ...)
}
\arguments{
\item{x}{a "LAM" object}

\item{y}{is ignored}

\item{...}{further arguments passed to and from other methods}
}
\description{
Plots a four panel graphic summairsing Load Apportionment Model fits.
}
\details{
\code{y} is included for consistency with generic \code{print} but is
     not used
}
\examples{
data(wensum)
     Q <- wensum$Q
     Co <- wensum$TP

     LAMwensum <- LAM(Q, Co)

     LAMplot(LAMwensum)
}

