% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_build_log.R
\name{load_build_log}
\alias{load_build_log}
\title{Load the ELS G2 Build List}
\usage{
load_build_log(path)
}
\arguments{
\item{path}{A path to the build log}
}
\value{
A data frame
}
\description{
Load the ELS G2 Build List
}
