% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\docType{package}
\name{loadr}
\alias{loadr}
\title{Cleaner R workspaces.}
\description{
Functions for loading data into a shared variable environment
}
\references{
\url{https://github.com/databio/loadr}
}
\author{
Nathan Sheffield
}
