% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadr.R
\name{loadr}
\alias{loadr}
\title{Attach several packages}
\usage{
loadr(..., .x = NULL, quiet = TRUE)
}
\arguments{
\item{...}{package names quoted or not.}

\item{.x}{a vector or a list that is concatenated to \code{...}.}

\item{quiet}{If \code{FALSE}, the function prints output. Defaults to \code{TRUE}.}
}
\value{
NULL invisibly.
}
\description{
Attach several packages
}
